@@ -12,8 +12,8 @@
     "axios": "^0.26.0",
     "daisyui": "^2.2.0",
     "framer-motion": "^6.2.8",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
     "react-icons": "^4.3.1",
     "react-redux": "^7.2.6",
     "react-router-dom": "^6.2.1",
