@@ -1,15 +1,16 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import './index.css';
-import App from './App';
-import { Provider } from 'react-redux';
-import { store } from './app/store';
+import React from 'react'
+import * as ReactDOMClient from 'react-dom/client'
+import './index.css'
+import App from './App'
+import { Provider } from 'react-redux'
+import { store } from './app/store'
 
-ReactDOM.render(
-  <React.StrictMode>
-    <Provider store={store}>
-      <App />
-    </Provider>
-  </React.StrictMode>,
-  document.getElementById('root')
-);
+const container = document.getElementById('root')
+
+const root = ReactDOMClient.createRoot(container)
+
+root.render(
+  <Provider store={store}>
+    <App />
+  </Provider>
+)
